Separate BackendContainer into own module
