+++ programs/castle-vault/src/instructions/rebalance.rs
@@ -1,10 +1,8 @@
-use std::cmp::Ordering;
 use std::ops::{Deref, Index};
 
 use anchor_lang::prelude::*;
 use anchor_lang::solana_program::program_pack::{IsInitialized, Pack, Sealed};
 use port_anchor_adaptor::PortReserve;
-use solana_maths::Decimal;
 use solana_maths::{Rate, TryAdd, TryMul};
 use strum::IntoEnumIterator;
 
@@ -357,121 +355,3 @@ fn get_apr_chris(
         .iter()
         .try_fold(Rate::zero(), |acc, r| acc.try_add(*r))
 }
-
-impl BackendContainer<Allocation> {
-    pub fn try_from_weights(
-        rates: &BackendContainer<Rate>,
-        vault_value: u64,
-        slot: u64,
-    ) -> Result<Self, ProgramError> {
-        rates.try_apply(|_provider, rate| {
-            rate.try_mul(vault_value).and_then(|product| {
-                Decimal::from(product)
-                    .try_floor_u64()
-                    .map(|value| Allocation {
-                        value,
-                        last_update: LastUpdate::new(slot),
-                    })
-            })
-        })
-    }
-}
-
-impl<T> BackendContainer<T>
-where
-    T: ReturnCalculator,
-{
-    fn compare(&self, lhs: &T, rhs: &T) -> Result<Ordering, ProgramError> {
-        Ok(lhs.calculate_return(0)?.cmp(&rhs.calculate_return(0)?))
-    }
-}
-
-impl BackendContainer<Reserves> {
-    fn calculate_weights_max_yield(&self) -> Result<BackendContainer<Rate>, ProgramError> {
-        self.into_iter()
-            .max_by(|(_prov_x, alloc_x), (_prov_y, alloc_y)| {
-                // TODO: can we remove the unwrap() in any way?
-                self.compare(*alloc_x, *alloc_y).unwrap()
-            })
-            .map(|(max_yielding_provider, _a)| {
-                self.apply(|provider, _v| {
-                    if provider == max_yielding_provider {
-                        Rate::one()
-                    } else {
-                        Rate::zero()
-                    }
-                })
-            })
-            // TODO make this error handling more granular and informative
-            .ok_or(ErrorCode::StrategyError)
-            .map_err(Into::into)
-    }
-
-    pub fn calculate_weights(
-        &self,
-        stype: StrategyType,
-    ) -> Result<BackendContainer<Rate>, ProgramError> {
-        match stype {
-            StrategyType::MaxYield => self.calculate_weights_max_yield(),
-            StrategyType::EqualAllocation => todo!(),
-        }
-    }
-}
-
-impl Strategy for BackendContainer<u16> {
-    fn calculate_weights(&self, _assets: &Assets) -> Result<StrategyWeights, ProgramError> {
-        todo!()
-    }
-
-    fn calculate_weights_chris<T>(&self) -> Result<BackendContainer<Rate>, ProgramError> {
-        todo!()
-    }
-
-    fn verify_weights_chris(&self) -> ProgramResult {
-        // let sum = Provider::iter()
-        //     .map(|p| proposed_weights[p])
-        //     .try_fold(Rate::zero(), |acc, x| acc.try_add(x))?;
-
-        // self.into_iter()
-        //     .try_fold(Rate::zero(), |acc, (_, x)| acc.try_add(x.value))?;
-        // .map(|(_provider, allocation)| {
-        //     if allocation.value > 0 {
-        //         Ok(())
-        //     } else {
-        //         Err(ErrorCode::RebalanceProofCheckFailed.into())
-        //     }
-        // })
-        // .collect::<Result<Vec<_>, ProgramError>>()?;
-
-        let sum = Rate::default();
-        if sum != Rate::one() {
-            return Err(ErrorCode::InvalidProposedWeights.into());
-        }
-        Ok(())
-    }
-}
-
-impl From<BackendContainer<u16>> for BackendContainer<Rate> {
-    fn from(c: BackendContainer<u16>) -> Self {
-        c.apply(|_provider, v| Rate::from_bips(*v as u64))
-    }
-}
-
-/*
-impl TryFrom<BackendContainer<Rate>> for BackendContainer<Allocation> {
-    type Error = ProgramError;
-
-    fn try_from(value: BackendContainer<Rate>) -> Result<Self, Self::Error> {
-        value.try_apply(|_provider, rate| {
-            rate.try_mul(vault_value).and_then(|product| {
-                Decimal::from(product)
-                    .try_floor_u64()
-                    .map(|value| Allocation {
-                        value,
-                        last_update: LastUpdate::new(clock.slot),
-                    })
-            })
-        })
-    }
-}
-*/
